% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Calibrates mzXML list object based on the calibrant peaks which is injected in the beginning (or at the end of) each chromatographic run.}
\usage{
calibration(
  mzXML = read.mzXML("C:/mymzXMLfolder/sample1.mzXML"),
  calibration_region = c(0.08, 0.26),
  calibrant_substance = "Na_Formate_pos",
  int_thres = 1000,
  mzthres = 0.05
)
}
\arguments{
\item{mzXML}{sample mzXML list created from read.mzXML function}

\item{calibration_region}{The retention time range in minutes, in which calibrant peaks appear}

\item{int_thres}{Intensity threshold above which calibrant peaks are taken into account}

\item{mzthres}{Threshold indicating the absolute difference of experimental and theoretical mass. If more than this value, the calibrant peak is disregarded.}

\item{calibration_substance}{It's the calibrant substance, based on which reference mass peaks are calculated.}
}
\value{
calibrated mzXML list object in the global environment
}
\description{
Calibrates mzXML list object and returns a calibrated mzXML list on the global environment variables. Then this calibrated mzXML list can be stored as mzXML file using function write.mzXML
}
\details{
This function isolates the calibrant peaks, which appear in the beginning of each chromatographic run. Calculates
the absolute error in mass Based on the reference masses of the calibrant. After that, an optimized polyonimal model is fitted
to the values (error=f(masses)). Based on this expression, all the masses in all the full scans are replaced with the predicted
values. 
At the moment calibration_substance argument can accept 16 calibrant substances;
\itemize{
 \item{"Arginine_neg"}
 \item{"Arginine_pos"}
 \item{"Cs_Perfluoroheptanoate_pos"}
 \item{"Fatty_Acids_neg"}
 \item{"Fatty_Acids_pos"}
 \item{"Li_Formate_neg"}
 \item{"Li_Formate_pos"}
 \item{"Na_Acetate_neg"}
 \item{"Na_Acetate_pos"}
 \item{"Na_Formate_neg"}
 \item{"Na_Formate_pos"}
 \item{"Na_TFA_neg"}
 \item{"Na_TFA_pos"}
 \item{"PEG_pos"}
 \item{"Reserpine_pos"}
 \item{"silicone_pos"}
}

In case another substance is used as calibrant, calibrant peaks can be inserted manually. For example the following
format should be followed, so that calibrant peaks can inserted manually:

calibration_substance=data.frame(substance=c("Peak1", "Peak2", "Peak3","Peak4","Peak5","Peak6","Peak7"),
                                refmz=c(90.9766,158.9641,226.9515,294.9389,362.9263,430.9138,498.9012))
}
\examples{
mzXMLexample<-read.mzXML(paste(list.files(file.path(find.package("peakTrams"), "data"),pattern=".mzXML",full.names = T)))
mzXMLexample_calibrated<-calibration(mzXML=mzXMLexample,calibration_region=c(0.08,0.26), calibrant_substance="Na_Formate_pos",int_thres=1000, mzthres=0.02)

}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
