% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRPHILIC.R
\name{compareRPHILIC}
\alias{compareRPHILIC}
\title{Performs peak picking based on centWave algorithm and compares peak list in two chromatographies}
\usage{
compareRPHILIC(
  HILIC = paste(getwd(), "Subtracted_HILIC.mzXML", sep = "/"),
  RP = paste(getwd(), "Subtracted_RP.mzXML", sep = "/"),
  mzthr = 0.01,
  doubleckeck = TRUE,
  ...
)
}
\arguments{
\item{HILIC}{path where HILIC mzXML file is stored}

\item{RP}{path where RP mzXML file is stored}

\item{mzthr}{absolute mass accuracy error}

\item{doubleckeck}{TRUE or FALSE value as described in Description section}

\item{...}{parameters of centWave peak picking algorithm (ppm, peakwidth, sn, prefilter, ...)}
}
\value{
Returns a list of 3 elements; common peaks, peaks unique in HILIC, peaks unique in RP
}
\description{
In the first step it performs peak picking on both HILIC and RP files according to centWave approach (See References for more details).
Afterwards, peak tables are splited on lists objects so that each element of the lists consists of chromatographic peaks clustered according to mass accuracy (threshold argument mzthr).
If doublecheck is set to TRUE; For each element of the list, mean mz is passed again to centWave as Region Of Interest (ROI) and in case extra peaks are detected they are added to the lists
After this Verifycation of peak picking, two columns entitled "numlist" and "chromatogr" are added. 
Numlist indicates the element number of the list (mz slice number) and "chromatogr" indicates chromatography which is either RP or HILIC.
Then, elements of both lists are compared and thus common peaks are detected and removed from the initial lists. Now initial lists are unique peaks in the chromatographies.
Again if doublecheck is set to TRUE; Unique peaks are passed as ROIs in centWave and if a pair peak is detected in the other chromatography,
the pair is moved to common peaks and is deleted from unique lists.
}
\references{
\url{http://msbi.ipb-halle.de/msbi/centwave} and DOI: 10.1186/1471-2105-9-504
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
