% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMS2.R
\name{extractMS2}
\alias{extractMS2}
\title{Extracts and saves MS/MS spectra from a directory with mzXML files}
\usage{
extractMS2(mzXML)
}
\arguments{
\item{sample}{mzXML list produced by read.mzXML function}
}
\value{
sample mzXML list without empty full scans
}
\description{
It is possible (especially in negative ionization) some MS/MS spectrum to be empty (without any spectral peaks).
This function removes the reduntant full scans and returns a mzXML object without the empty full scans.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
