% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getinfo}
\alias{getinfo}
\title{Gets retention time and number of peaks of full scans of a mzXML list}
\usage{
getinfo(sample)
}
\arguments{
\item{sample}{mzXML list created from read.mzXML function}
}
\value{
A data frame with number of scan, with retention time of each full scan, mslevel, number of spectral peaks and in case
of MS/MS full scan precursor mass and precurson intensity.
}
\description{
Takes in a raw sample and returns a data frame with retention time of each full scan
}
\examples{
sample_mzXML<-read.mzXML(list.files(paste(find.package(package="peakTrams"),"data",sep="/"),pattern = ".mzXML", full.names = TRUE))
getrt(sample=sample_mzXML)

}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
