% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXIC.R
\name{plotXIC}
\alias{plotXIC}
\title{Performs eΧtracted Ιon Chromatogram (XIC)}
\usage{
plotXIC(
  sample = sub_RP,
  mzrange = c(272.0275),
  accuracy = 0.005,
  timerange = c(round(min(getinfo(sample)$timeofscan/60), 1),
    round(max(getinfo(sample)$timeofscan/60), 1)),
  ploteic = TRUE,
  extra_title = "EIC of",
  ...
)
}
\arguments{
\item{sample}{mzXML list object (e.g. read.mzXML("c:/myfolder/sample.mzXML")}

\item{mzrange}{mz}

\item{accuracy}{absolute mass accuracy}

\item{timerange}{time range}

\item{ploteic}{If TRUE plots extracted ion chromatogram otherwise just returns the dataframe object without plotting}

\item{extra_title}{Extra title in the beginning of the main title in figure}

\item{...}{other arguments passed to plot function (such as type for type of line,col for color of line etc)}

\item{type}{what type of plot should be drawn. Possible types are}
}
\value{
A dataframe with three columns mz,intensity and timeofscan
}
\description{
Performs extracted ion chromatogram using specific mass accuracy window and optionally retention time window. 
In case mzrange is a vector with one element, accuracy argument should be specified, otherwise mzrange should be a vector 
of two elements. 
Extra title can be a vector with two elements. The first element is added in the beginning of the title and the second at the end of
the title.
}
\examples{
plotXIC(sample=paste(list.files(file.path(find.package("peakTrams"), "data"),pattern=c("Sample_RP",".mzXML"),full.names = T)), mzrange=c(286.1438), accuracy=0.005, type="l")
plotXIC(sample=paste(list.files(file.path(find.package("peakTrams"), "data"),pattern=c("Sample_RP",".mzXML"),full.names = T)), mzrange=c(286.1438), accuracy=0.005, type="l", col="red", timerange=c(2,3.5))
plotXIC(sample=paste(list.files(file.path(find.package("peakTrams"), "data"),pattern=c("Sample_RP",".mzXML"),full.names = T)), mzrange=c(286.1438), accuracy=0.005, type="l", col="red", timerange=c(2,3.5), extra_title=c("EIC of","Morphine [M+H]"))
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
