% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{prioritization}
\alias{prioritization}
\title{Prioritizes unique and common peak lists.}
\usage{
prioritization(output, deadvolume = c(1.5))
}
\arguments{
\item{output}{object returned from compare function.}

\item{deadvolume}{dead volume of HILIC column}
}
\value{
Returns a list with two data frame objects, which contain the prioritized common and unique peaks respectively.
}
\description{
Unique peaks are prioritized based on the observed maximum intensity. 
Common peaks are prioritized based on three criteria. 
The first criterio is that peaks should be retained in HILIC chromatography. 
This happens only if compounds elutes after the dead volume of the column, which can be set 
from deadvolume argument.
The second criterio is is related to the fact that we are looking for polar compounds. Such compounds
are eluted relatively in the beginning of Reversed-phase chromatography, whether are eluted later in HILIC
chromatography. So, the second requirement is that retention time in HILIC is higher than in RP.
The third criterio states that polar compounds should give better response in HILIC than in RP.
}
\details{
According to the 3 criteria, common peaks are divided into 8 categories, which were eliminated to 5
because 3 categories have no analytical explanation and normally consist of very few common pair peaks.

The categories are the followings:
\itemize{
 \item{"Category 5"}{ All the criteria are valid. Therefore, these are the peaks of interest}
 \item{"Category 4"}{ Criteria 1 and 2 are valid. Still some compounds of interest may be in this group}
 \item{"Category 3"}{ Criteria 1 and 3 are valid.}
 \item{"Category 2"}{ Criterio 1 is valid.}
 \item{"Category 1"}{ Criterio 1 is invalid.}
}
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
