% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mzXML.R
\name{read.mzXML}
\alias{read.mzXML}
\title{This function reads mzXML files}
\usage{
read.mzXML(filename)
}
\arguments{
\item{filename}{The directory in the hard drive that the mzXML files is stored.}
}
\value{
Returns a list object, which contains the following elements;
\item{header}{Stores header of <mzXML> section containing information about namespace and schema file location.}
\item{parentFile}{Path to all the ancestor files. Stored as XML.}
\item{dataProcessing}{Description of any data manipulation. Stored as XML.}
\item{msInstrument}{General information about the MS instrument. Stored as XML.}
\item{scan}{List of Mass Spectra scans. Each element of the list contain the following elements;}
\item{peaks}{ peak intensities of the scan}
\item{mass}{ masses (m/z) corresponding to \code{peaks}. Vectors \code{mass} and \code{peaks} have the same length.}
\item{num}{ scan number}
\item{parentNum}{ scan number of parent scan in case of recursively stored scans (\code{msLevel>1})}
\item{msLevel}{ Level 1 means MS1, while level 2 means MS2, etc.}
\item{scanAttr}{ Other useful information, such as retention time, polarity, collision energy, total ion current}
\item{maldi}{ acquisition dependent properties of a MALDI experiment (optional)}
\item{scanOrigin}{ name of parent file}
\item{precursorMz}{ information about the precursor ion}
\item{nameValue}{ properties of the scan not included elsewhere}
}
\description{
Reads a mzXML file and returns a mzXML list object in the global environment.
}
\details{
This functions reads a mzXML file and stores it as a list in the variables global environment.
}
\examples{
#Donot run
library("peakTrams")
sample<-read.mzXML(filename=c:\R_working_directory\sample.mzXML)

}
\references{
Definition of \code{mzXML} format:
\url{http://tools.proteomecenter.org/mzXMLschema.php}

Documentation of \code{mzXML} format:
\url{http://sashimi.sourceforge.net/schema_revision/mzXML_2.1/Doc/mzXML_2.1_tutorial.pdf}

More Documentation of \code{mzXML} format:
 \url{http://sashimi.sourceforge.net/software_glossolalia.html}
}
\author{
Originally this function was part of caMassClass R-package by Jarek Tuszynski

Codes maintained by Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
