% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEMPTYscans.R
\name{removeEMPTYscans}
\alias{removeEMPTYscans}
\title{Removes empty scan events from a mzXML file}
\usage{
removeEMPTYscans(sample)
}
\arguments{
\item{sample}{mzXML list produced by read.mzXML function}
}
\value{
new_sample mzXML list without empty full scans
}
\description{
It is possible (especially in negative ionization) some MS/MS spectrum to be empty (without any spectral peaks).
This function removes the reduntant full scans and returns a mzXML object without the empty full scans.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
