% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMS2.R
\name{removeMS2}
\alias{removeMS2}
\title{Removes all MS2 scan events from a mzXML list}
\usage{
removeMS2(sample)
}
\arguments{
\item{sample}{mzXML list object}
}
\value{
sample mzXML list object without MS/MS spectra
}
\description{
Takes as input an object which was created by read.mzXML function 
and returns an object without the MS/MS full scans.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
