% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMSn.R
\name{removeMSn}
\alias{removeMSn}
\title{Removes all MSn scan events from a mzXML list}
\usage{
removeMSn(sample)
}
\arguments{
\item{sample}{mzXML list object}
}
\value{
sample mzXML list object without MS/MS spectra
}
\description{
Takes as input an object which was created by read.mzXML function 
and returns an object without the MSn scans.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
