% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removenoise.R
\name{removenoise}
\alias{removenoise}
\title{Removes noisy spectral peaks below an intensity threshold in all full scans}
\usage{
removenoise(
  mzXML = read.mzXML(list.files(paste(find.package(package = "peakTrams"), "data", sep
    = "/"), pattern = ".mzXML", full.names = TRUE)),
  noisethreshold = 100,
  mslevel = c(1)
)
}
\arguments{
\item{mzXML}{mzXML list produced by read.mzXML function}

\item{noisethreshold}{noise threshold}

\item{mslevel}{ms level, can be c(1) or c(2) or c(1,2)}
}
\value{
Returns a new mzXML list without spectral peaks below noisethreshold
}
\description{
The function goes in all scans of a specific mslevel, reads the spectra and removes peaks below noisethreshold.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
