% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removescans.R
\name{removescans}
\alias{removescans}
\title{Removes selected full scans from a mzXML list object}
\usage{
removescans(mzXML = blank_HILIC, scansORtime = c(17, 25), time = TRUE)
}
\arguments{
\item{mzXML}{file produced from read.mzXML function}

\item{scansORtime}{Selected scans (or scans with retention time if time=TRUE) to be removed}

\item{time}{Logical value TRUE or FALSE}
}
\value{
a mzXML list without selected full scans
}
\description{
Takes as input an object which was created by read.mzXML function 
and returns an object without selected scans passed in scansORtime argument.
In case time is set to TRUE then scanORtime should be a vector of two elements containing
retention time in minutes. The selected full scans with retention time within this interval will
be deleted from the mzXML list.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
