% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMS2.R
\name{searchMS2}
\alias{searchMS2}
\title{It searches MS/MS spectra for a mass within mass and retention time tolerance}
\usage{
searchMS2(
  sample,
  mass = 100.0323,
  tolerancemass = 0.05,
  rt = 3,
  tolerancert = 0.5
)
}
\arguments{
\item{sample}{mzXML list produced by read.mzXML}

\item{mass}{selected mass}

\item{tolerancemass}{absolute mass accuracy set as tolerance value}

\item{rt}{retention time in minutes}

\item{tolerancert}{tolerance in retention time (given in minutes)}
}
\value{
Peak list data frame (mz and intensity) with MS/MS spectra.
}
\description{
The function searches MS/MS full scans and looks for the selected mass within the given
tolerance mass. The MS/MS spectrum should be within a given retention time window, as this
is specified in rt and tolerancert arguments. In case many MS/MS spectrum exist in the file, the one with the most peak is selected.
}
\examples{
sample_mzXML<-read.mzXML(list.files(paste(find.package(package="peakTrams"),"data",sep="/"),pattern = ".mzXML", full.names = TRUE))
searchMS2(sample=sample_mzXML, mass=100.0323, tolerancemass=0.05, rt=3.00, tolerancert=0.5)

}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
