% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtracts procedural blank from sample chromatogram}
\usage{
subtract(
  sample = read.mzXML(paste(list.files(file.path(find.package("peakTrams"), "data"),
    pattern = c("Sample_RP", ".mzXML"), full.names = T))),
  blank = read.mzXML(paste(list.files(file.path(find.package("peakTrams"), "data"),
    pattern = c("Blank_RP", ".mzXML"), full.names = T))),
  mzdiff = 0.01,
  filter = 100
)
}
\arguments{
\item{sample}{sample mzXML list (provided from read.mzXML function)}

\item{blank}{blank mzXML list (provided from read.mzXML function)}

\item{mzdiff}{Absolute mass accuracy threshold}

\item{filter}{cut off filter (normally is set at a small intensity value)}
}
\value{
an mzXML list object containg subtracted chromatogram
}
\description{
This function takes into account only MS1 levels (other levels are removed). 
It goes to all scans of sample, searches for the nearest in retention time full scan in blank and 
then reads the spectral peaks. Aftewards, checks one by one the spectral peaks in sample and 
searches for possible match in blank withing a given mass accuracy threshold (see argument mzdiff). 
In case a spectral peak is present in blank then the intensity of the sample is subtracted by the 
intenisity observed in blank. In case the result is a negative intensity value, it is converted to zero.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
