% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mzXML.R
\name{write.mzXML}
\alias{write.mzXML}
\title{Returns mzXML list object in mzXML files}
\usage{
write.mzXML(mzXML, filename, precision = c("32", "64"))
}
\arguments{
\item{mzXML}{mzXML list created from read.mzXML}

\item{filename}{The directory in the hard drive that the mzXML files will be stored}

\item{precision}{option between 32 bit or 64 bit precision}
}
\value{
Returns an mzXML file back to hard drive
}
\description{
Converts the mzXML list object to a mzXML file and stores it on the hard drive
}
\author{
Originally this function was part of caMassClass R-package by Jarek Tuszynski

Codes maintained by Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
