% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeoutput.R
\name{writeoutput}
\alias{writeoutput}
\title{Performs Extracted Ion Chromatograms (EICs) for unique peaks in RP, unique peaks in HILIC, and common peaks of category 5.}
\usage{
writeoutput(
  prioritized_output = prioritized_output,
  sampleRP = paste(getwd(), "Subtracted_RP.mzXML", sep = "/"),
  sampleHILIC = paste(getwd(), "Subtracted_HILIC.mzXML", sep = "/"),
  accuracy = 0.005,
  numofcommon = length(prioritized_output[[1]][prioritized_output[[1]]$category == 5,
    1]),
  numofuniqueHILIC = length(prioritized_output[[2]][, 1]),
  numofuniqueRP = length(prioritized_output[[3]][, 1]),
  ...
)
}
\arguments{
\item{prioritized_output}{output created by prioritization function}

\item{sampleRP}{mzXML list object of chromatograms in RP}

\item{sampleHILIC}{mzXML list object of chromatograms in HILIC}

\item{accuracy}{absolute mass accuracy}

\item{numofcommon}{number of common features (from category 5) that user wants to be plotted}

\item{numofuniqueHILIC}{number of HILIC unique features that user wants to be plotted}

\item{numofuniqueRP}{number of RP unique features that user wants to be plotted}
}
\value{
Created 3 csv files with common peaks, unique peaks in HILIC and 
unique peaks in RP and 3 pdf files with the EICs.
}
\description{
Plots the prioritized output and saves in pdf files the EICs. User should specify the path of the mzXML files. In case numofcommon,
numofuniqueHILIC and numofuniqueRP are not passed then all the EICs will be plotted, which may lead to high waiting time.
}
\author{
Nikiforos Alygizakis <nalygizakis@chem.uoa.gr>
}
